Buffs, killer setting, new test level
Now when the player gets all the transformation itens, it gets a buff and get the killer status. Made a smal testing levels with all the platform types i've made so far to test it out. Seens fun, but I need my jam partner to wake up to confirm it